@@ -5,10 +5,11 @@ import { AuthGuard } from './shared';
 const routes: Routes = [
     {
         path: '',
-        loadChildren: () => import('./layout/layout.module').then((m) => m.LayoutModule),
+        loadChildren: () => import('./pages/layout.module').then((m) => m.LayoutModule),
         canActivate: [AuthGuard]
     },
     { path: 'login', loadChildren: () => import('./login/login.module').then((m) => m.LoginModule) },
+    { path: 'auth-callback', loadChildren: () => import('./auth-callback/auth-callback.module').then((m) => m.AuthCallbackModule) },
     { path: 'signup', loadChildren: () => import('./signup/signup.module').then((m) => m.SignupModule) },
     {
         path: 'error',
@@ -26,4 +27,4 @@ const routes: Routes = [
     imports: [RouterModule.forRoot(routes)],
     exports: [RouterModule]
 })
-export class AppRoutingModule {}
+export class AppRoutingModule { }
