@@ -1,6 +1,9 @@
 import { Component, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
 import { routerTransition } from '../router.animations';
+import { NgxSpinner } from 'ngx-spinner/lib/ngx-spinner.enum';
+import { AuthService } from '../shared/services';
+import { NgxSpinnerService } from 'ngx-spinner';
 
 @Component({
     selector: 'app-login',
@@ -9,11 +12,17 @@ import { routerTransition } from '../router.animations';
     animations: [routerTransition()]
 })
 export class LoginComponent implements OnInit {
-    constructor(public router: Router) {}
+    constructor(
+        // public router: Router,
+        private spinner: NgxSpinnerService,
+        private authService: AuthService
+    ) { }
 
-    ngOnInit() {}
+    ngOnInit() { }
 
     onLoggedin() {
-        localStorage.setItem('isLoggedin', 'true');
+        // localStorage.setItem('isLoggedin', 'true');
+        this.spinner.show();
+        this.authService.login();
     }
 }
