@@ -1,12 +1,19 @@
 import { CommonModule } from '@angular/common';
-import { NgModule } from '@angular/core';
+import { NgModule, CUSTOM_ELEMENTS_SCHEMA } from '@angular/core';
 import { TranslateModule } from '@ngx-translate/core';
 
 import { LoginRoutingModule } from './login-routing.module';
 import { LoginComponent } from './login.component';
+import { NgxSpinnerModule } from 'ngx-spinner';
 
 @NgModule({
-    imports: [CommonModule, TranslateModule, LoginRoutingModule],
-    declarations: [LoginComponent]
+    imports: [
+        CommonModule,
+        TranslateModule,
+        LoginRoutingModule,
+        NgxSpinnerModule
+    ],
+    declarations: [LoginComponent],
+    schemas: [CUSTOM_ELEMENTS_SCHEMA]
 })
-export class LoginModule {}
+export class LoginModule { }
