@@ -3,11 +3,13 @@ using KnowledgeSpace.BackendServer.Constants;
 using KnowledgeSpace.BackendServer.Data;
 using KnowledgeSpace.BackendServer.Data.Entities;
 using KnowledgeSpace.BackendServer.Helpers;
+using KnowledgeSpace.BackendServer.Services;
 using KnowledgeSpace.ViewModels;
 using KnowledgeSpace.ViewModels.Contents;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
+using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 
@@ -16,10 +18,13 @@ namespace KnowledgeSpace.BackendServer.Controllers
     public class CategoriesController : BaseController
     {
         private readonly ApplicationDbContext _context;
+        private readonly ICacheService _cacheService;
 
-        public CategoriesController(ApplicationDbContext context)
+        public CategoriesController(ApplicationDbContext context,
+            ICacheService cacheService)
         {
             _context = context;
+            _cacheService = cacheService;
         }
 
         [HttpPost]
@@ -40,6 +45,8 @@ namespace KnowledgeSpace.BackendServer.Controllers
 
             if (result > 0)
             {
+                await _cacheService.RemoveAsync("Categories");
+
                 return CreatedAtAction(nameof(GetById), new { id = category.Id }, request);
             }
             else
@@ -52,11 +59,17 @@ namespace KnowledgeSpace.BackendServer.Controllers
         [AllowAnonymous]
         public async Task<IActionResult> GetCategories()
         {
-            var categorys = await _context.Categories.ToListAsync();
+            var cachedData = await _cacheService.GetAsync<List<CategoryVm>>("Categories");
+            if (cachedData == null)
+            {
+                var categorys = await _context.Categories.ToListAsync();
 
-            var categoryvms = categorys.Select(c => CreateCategoryVm(c)).ToList();
+                var categoryVms = categorys.Select(c => CreateCategoryVm(c)).ToList();
+                await _cacheService.SetAsync("Categories", categoryVms);
+                cachedData = categoryVms;
+            }
 
-            return Ok(categoryvms);
+            return Ok(cachedData);
         }
 
         [HttpGet("filter")]
@@ -121,6 +134,8 @@ namespace KnowledgeSpace.BackendServer.Controllers
 
             if (result > 0)
             {
+                await _cacheService.RemoveAsync("Categories");
+
                 return NoContent();
             }
             return BadRequest(new ApiBadRequestResponse("Update category failed"));
@@ -138,6 +153,8 @@ namespace KnowledgeSpace.BackendServer.Controllers
             var result = await _context.SaveChangesAsync();
             if (result > 0)
             {
+                await _cacheService.RemoveAsync("Categories");
+
                 CategoryVm categoryvm = CreateCategoryVm(category);
                 return Ok(categoryvm);
             }
