@@ -9,10 +9,12 @@ using KnowledgeSpace.ViewModels;
 using KnowledgeSpace.ViewModels.Contents;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
+using Microsoft.AspNetCore.Identity.UI.Services;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Logging;
 using System;
+using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using System.Net.Http.Headers;
@@ -26,16 +28,28 @@ namespace KnowledgeSpace.BackendServer.Controllers
         private readonly ISequenceService _sequenceService;
         private readonly IStorageService _storageService;
         private readonly ILogger<KnowledgeBasesController> _logger;
+        private readonly IEmailSender _emailSender;
+        private readonly IViewRenderService _viewRenderService;
+        private readonly ICacheService _cacheService;
+        private readonly IOneSignalService _oneSignalService;
 
         public KnowledgeBasesController(ApplicationDbContext context,
             ISequenceService sequenceService,
             IStorageService storageService,
-            ILogger<KnowledgeBasesController> logger)
+            ILogger<KnowledgeBasesController> logger,
+            IEmailSender emailSender,
+            IViewRenderService viewRenderService,
+            ICacheService cacheService,
+            IOneSignalService oneSignalService)
         {
             _context = context;
             _sequenceService = sequenceService;
             _storageService = storageService;
             _logger = logger ?? throw new ArgumentNullException(nameof(logger));
+            _emailSender = emailSender;
+            _viewRenderService = viewRenderService;
+            _cacheService = cacheService;
+            _oneSignalService = oneSignalService;
         }
 
         [HttpPost]
@@ -74,7 +88,13 @@ namespace KnowledgeSpace.BackendServer.Controllers
 
             if (result > 0)
             {
+                await _cacheService.RemoveAsync(CacheConstants.LatestKnowledgeBases);
+                await _cacheService.RemoveAsync(CacheConstants.PopularKnowledgeBases);
+
                 _logger.LogInformation("End PostKnowledgeBase API - Success");
+
+                await _oneSignalService.SendAsync(request.Title, request.Description,
+                     string.Format(CommonConstants.KnowledgeBaseUrl, knowledgeBase.SeoAlias, knowledgeBase.Id));
 
                 return CreatedAtAction(nameof(GetById), new
                 {
@@ -111,52 +131,64 @@ namespace KnowledgeSpace.BackendServer.Controllers
         [AllowAnonymous]
         public async Task<IActionResult> GetLatestKnowledgeBases(int take)
         {
-            var knowledgeBases = from k in _context.KnowledgeBases
-                                 join c in _context.Categories on k.CategoryId equals c.Id
-                                 orderby k.CreateDate descending
-                                 select new { k, c };
+            var cachedData = await _cacheService.GetAsync<List<KnowledgeBaseQuickVm>>(CacheConstants.LatestKnowledgeBases);
+            if (cachedData == null)
+            {
+                var knowledgeBases = from k in _context.KnowledgeBases
+                                     join c in _context.Categories on k.CategoryId equals c.Id
+                                     orderby k.CreateDate descending
+                                     select new { k, c };
 
-            var knowledgeBasevms = await knowledgeBases.Take(take)
-                .Select(u => new KnowledgeBaseQuickVm()
-                {
-                    Id = u.k.Id,
-                    CategoryId = u.k.CategoryId,
-                    Description = u.k.Description,
-                    SeoAlias = u.k.SeoAlias,
-                    Title = u.k.Title,
-                    CategoryAlias = u.c.SeoAlias,
-                    CategoryName = u.c.Name,
-                    NumberOfVotes = u.k.NumberOfVotes,
-                    CreateDate = u.k.CreateDate
-                }).ToListAsync();
+                var knowledgeBaseVms = await knowledgeBases.Take(take)
+                    .Select(u => new KnowledgeBaseQuickVm()
+                    {
+                        Id = u.k.Id,
+                        CategoryId = u.k.CategoryId,
+                        Description = u.k.Description,
+                        SeoAlias = u.k.SeoAlias,
+                        Title = u.k.Title,
+                        CategoryAlias = u.c.SeoAlias,
+                        CategoryName = u.c.Name,
+                        NumberOfVotes = u.k.NumberOfVotes,
+                        CreateDate = u.k.CreateDate
+                    }).ToListAsync();
+                await _cacheService.SetAsync(CacheConstants.LatestKnowledgeBases, knowledgeBaseVms, 2);
+                cachedData = knowledgeBaseVms;
+            }
 
-            return Ok(knowledgeBasevms);
+            return Ok(cachedData);
         }
 
         [HttpGet("popular/{take:int}")]
         [AllowAnonymous]
         public async Task<IActionResult> GetPopularKnowledgeBases(int take)
         {
-            var knowledgeBases = from k in _context.KnowledgeBases
-                                 join c in _context.Categories on k.CategoryId equals c.Id
-                                 orderby k.ViewCount descending
-                                 select new { k, c };
+            var cachedData = await _cacheService.GetAsync<List<KnowledgeBaseQuickVm>>(CacheConstants.PopularKnowledgeBases);
+            if (cachedData == null)
+            {
+                var knowledgeBases = from k in _context.KnowledgeBases
+                                     join c in _context.Categories on k.CategoryId equals c.Id
+                                     orderby k.ViewCount descending
+                                     select new { k, c };
 
-            var knowledgeBasevms = await knowledgeBases.Take(take)
-                .Select(u => new KnowledgeBaseQuickVm()
-                {
-                    Id = u.k.Id,
-                    CategoryId = u.k.CategoryId,
-                    Description = u.k.Description,
-                    SeoAlias = u.k.SeoAlias,
-                    Title = u.k.Title,
-                    CategoryAlias = u.c.SeoAlias,
-                    CategoryName = u.c.Name,
-                    NumberOfVotes = u.k.NumberOfVotes,
-                    CreateDate = u.k.CreateDate
-                }).ToListAsync();
+                var knowledgeBaseVms = await knowledgeBases.Take(take)
+                    .Select(u => new KnowledgeBaseQuickVm()
+                    {
+                        Id = u.k.Id,
+                        CategoryId = u.k.CategoryId,
+                        Description = u.k.Description,
+                        SeoAlias = u.k.SeoAlias,
+                        Title = u.k.Title,
+                        CategoryAlias = u.c.SeoAlias,
+                        CategoryName = u.c.Name,
+                        NumberOfVotes = u.k.NumberOfVotes,
+                        CreateDate = u.k.CreateDate
+                    }).ToListAsync();
+                await _cacheService.SetAsync(CacheConstants.PopularKnowledgeBases, knowledgeBaseVms, 24);
+                cachedData = knowledgeBaseVms;
+            }
 
-            return Ok(knowledgeBasevms);
+            return Ok(cachedData);
         }
 
         [HttpGet("filter")]
@@ -295,6 +327,8 @@ namespace KnowledgeSpace.BackendServer.Controllers
 
             if (result > 0)
             {
+                await _cacheService.RemoveAsync("LatestKnowledgeBases");
+                await _cacheService.RemoveAsync("PopularKnowledgeBases");
                 return NoContent();
             }
             return BadRequest(new ApiBadRequestResponse($"Update knowledge base failed"));
@@ -312,6 +346,9 @@ namespace KnowledgeSpace.BackendServer.Controllers
             var result = await _context.SaveChangesAsync();
             if (result > 0)
             {
+                await _cacheService.RemoveAsync(CacheConstants.LatestKnowledgeBases);
+                await _cacheService.RemoveAsync(CacheConstants.PopularKnowledgeBases);
+
                 KnowledgeBaseVm knowledgeBasevm = CreateKnowledgeBaseVm(knowledgeBase);
                 return Ok(knowledgeBasevm);
             }
@@ -335,6 +372,26 @@ namespace KnowledgeSpace.BackendServer.Controllers
             }).ToListAsync();
 
             return Ok(labels);
+        }
+
+        [HttpPut("{id}/view-count")]
+        [AllowAnonymous]
+        public async Task<IActionResult> UpdateViewCount(int id)
+        {
+            var knowledgeBase = await _context.KnowledgeBases.FindAsync(id);
+            if (knowledgeBase == null)
+                return NotFound();
+            if (knowledgeBase.ViewCount == null)
+                knowledgeBase.ViewCount = 0;
+
+            knowledgeBase.ViewCount += 1;
+            _context.KnowledgeBases.Update(knowledgeBase);
+            var result = await _context.SaveChangesAsync();
+            if (result > 0)
+            {
+                return Ok();
+            }
+            return BadRequest();
         }
 
         #region Private methods
@@ -461,7 +518,10 @@ namespace KnowledgeSpace.BackendServer.Controllers
 
             knowledgeBase.Title = request.Title;
 
-            knowledgeBase.SeoAlias = request.SeoAlias;
+            if (string.IsNullOrEmpty(request.SeoAlias))
+                knowledgeBase.SeoAlias = TextHelper.ToUnsignString(request.Title);
+            else
+                knowledgeBase.SeoAlias = request.SeoAlias;
 
             knowledgeBase.Description = request.Description;
 
@@ -477,7 +537,8 @@ namespace KnowledgeSpace.BackendServer.Controllers
 
             knowledgeBase.Note = request.Note;
 
-            knowledgeBase.Labels = string.Join(',', request.Labels);
+            if (request.Labels != null)
+                knowledgeBase.Labels = string.Join(',', request.Labels);
         }
 
         #endregion Private methods
