@@ -2,8 +2,10 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using KnowledgeSpace.BackendServer.Constants;
 using KnowledgeSpace.BackendServer.Data;
 using KnowledgeSpace.BackendServer.Helpers;
+using KnowledgeSpace.BackendServer.Services;
 using KnowledgeSpace.ViewModels.Contents;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
@@ -15,9 +17,11 @@ namespace KnowledgeSpace.BackendServer.Controllers
     public class LabelsController : BaseController
     {
         private readonly ApplicationDbContext _context;
+        private readonly ICacheService _cacheService;
 
-        public LabelsController(ApplicationDbContext context)
+        public LabelsController(ApplicationDbContext context, ICacheService cacheService)
         {
+            _cacheService = cacheService;
             _context = context;
         }
 
@@ -42,23 +46,29 @@ namespace KnowledgeSpace.BackendServer.Controllers
         [AllowAnonymous]
         public async Task<List<LabelVm>> GetPopularLabels(int take)
         {
-            var query = from l in _context.Labels
-                        join lik in _context.LabelInKnowledgeBases on l.Id equals lik.LabelId
-                        group new { l.Id, l.Name } by new { l.Id, l.Name } into g
-                        select new
-                        {
-                            g.Key.Id,
-                            g.Key.Name,
-                            Count = g.Count()
-                        };
-            var labels = await query.OrderByDescending(x => x.Count).Take(take)
-                .Select(l => new LabelVm()
-                {
-                    Id = l.Id,
-                    Name = l.Name
-                }).ToListAsync();
+            var cachedData = await _cacheService.GetAsync<List<LabelVm>>(CacheConstants.PopularLabels);
+            if (cachedData == null)
+            {
+                var query = from l in _context.Labels
+                            join lik in _context.LabelInKnowledgeBases on l.Id equals lik.LabelId
+                            group new { l.Id, l.Name } by new { l.Id, l.Name } into g
+                            select new
+                            {
+                                g.Key.Id,
+                                g.Key.Name,
+                                Count = g.Count()
+                            };
+                var labels = await query.OrderByDescending(x => x.Count).Take(take)
+                    .Select(l => new LabelVm()
+                    {
+                        Id = l.Id,
+                        Name = l.Name
+                    }).ToListAsync();
+                await _cacheService.SetAsync(CacheConstants.PopularLabels, labels);
+                cachedData = labels;
+            }
 
-            return labels;
+            return cachedData;
         }
     }
 }
