@@ -1,6 +1,7 @@
 ﻿using System.Linq;
 using System.Threading.Tasks;
 using KnowledgeSpace.BackendServer.Data.Entities;
+using KnowledgeSpace.BackendServer.Extensions;
 using KnowledgeSpace.BackendServer.Helpers;
 using KnowledgeSpace.ViewModels.Contents;
 using Microsoft.AspNetCore.Mvc;
@@ -29,31 +30,38 @@ namespace KnowledgeSpace.BackendServer.Controllers
         }
 
         [HttpPost("{knowledgeBaseId}/votes")]
-        [ApiValidationFilter]
-        public async Task<IActionResult> PostVote(int knowledgeBaseId, [FromBody]VoteCreateRequest request)
+        public async Task<IActionResult> PostVote(int knowledgeBaseId)
         {
-            var vote = await _context.Votes.FindAsync(knowledgeBaseId, request.UserId);
-            if (vote != null)
-                return BadRequest(new ApiBadRequestResponse("This user has been voted for this KB"));
-
-            vote = new Vote()
-            {
-                KnowledgeBaseId = knowledgeBaseId,
-                UserId = request.UserId
-            };
-            _context.Votes.Add(vote);
-
+            var userId = User.GetUserId();
             var knowledgeBase = await _context.KnowledgeBases.FindAsync(knowledgeBaseId);
-            if (knowledgeBase != null)
+            if (knowledgeBase == null)
                 return BadRequest(new ApiBadRequestResponse($"Cannot found knowledge base with id {knowledgeBaseId}"));
 
-            knowledgeBase.NumberOfVotes = knowledgeBase.NumberOfVotes.GetValueOrDefault(0) + 1;
+            var numberOfVotes = await _context.Votes.CountAsync(x => x.KnowledgeBaseId == knowledgeBaseId);
+            var vote = await _context.Votes.FindAsync(knowledgeBaseId, userId);
+            if (vote != null)
+            {
+                _context.Votes.Remove(vote);
+                numberOfVotes -= 1;
+            }
+            else
+            {
+                vote = new Vote()
+                {
+                    KnowledgeBaseId = knowledgeBaseId,
+                    UserId = userId
+                };
+                _context.Votes.Add(vote);
+                numberOfVotes += 1;
+            }
+
+            knowledgeBase.NumberOfVotes = numberOfVotes;
             _context.KnowledgeBases.Update(knowledgeBase);
 
             var result = await _context.SaveChangesAsync();
             if (result > 0)
             {
-                return NoContent();
+                return Ok(numberOfVotes);
             }
             else
             {
