@@ -9,13 +9,14 @@ namespace KnowledgeSpace.ViewModels.Contents
     {
         public KnowledgeBaseCreateRequestValidator()
         {
-            RuleFor(x => x.CategoryId).GreaterThan(0).WithMessage("Category is required");
+            RuleFor(x => x.CategoryId).GreaterThan(0)
+                .WithMessage(string.Format(Messages.Required, "Danh mục"));
 
-            RuleFor(x => x.Title).NotEmpty().WithMessage("Title is required");
+            RuleFor(x => x.Title).NotEmpty().WithMessage(string.Format(Messages.Required, "Tiêu đề"));
 
-            RuleFor(x => x.Problem).NotEmpty().WithMessage("Problem is required");
+            RuleFor(x => x.Problem).NotEmpty().WithMessage(string.Format(Messages.Required, "Vấn đề"));
 
-            RuleFor(x => x.Note).NotEmpty().WithMessage("Note is required");
+            RuleFor(x => x.Note).NotEmpty().WithMessage(string.Format(Messages.Required, "Giải pháp"));
         }
     }
 }
