@@ -8,10 +8,7 @@ namespace KnowledgeSpace.ViewModels.Contents
     {
         public int? KnowledgeBaseId { get; set; }
 
-        public int? CommentId { get; set; }
-
         public string Content { get; set; }
-
-        public string ReportUserId { get; set; }
+        public string CaptchaCode { get; set; }
     }
 }
