@@ -4,6 +4,7 @@ using System.Linq;
 using System.Threading.Tasks;
 using KnowledgeSpace.ViewModels.Contents;
 using KnowledgeSpace.WebPortal.Extensions;
+using KnowledgeSpace.WebPortal.Helpers;
 using KnowledgeSpace.WebPortal.Models;
 using KnowledgeSpace.WebPortal.Services;
 using Microsoft.AspNetCore.Mvc;
@@ -75,32 +76,75 @@ namespace KnowledgeSpace.WebPortal.Controllers
             var knowledgeBase = await _knowledgeBaseApiClient.GetKnowledgeBaseDetail(id);
             var category = await _categoryApiClient.GetCategoryById(knowledgeBase.CategoryId);
             var labels = await _knowledgeBaseApiClient.GetLabelsByKnowledgeBaseId(id);
-            var user = await _userApiClient.GetById(User.GetUserId());
+
             var viewModel = new KnowledgeBaseDetailViewModel()
             {
                 Detail = knowledgeBase,
                 Category = category,
-                Labels = labels,
-                CurrentUser = user
+                Labels = labels
             };
+            if (User.Identity.IsAuthenticated)
+            {
+                viewModel.CurrentUser = await _userApiClient.GetById(User.GetUserId());
+            }
+            await _knowledgeBaseApiClient.UpdateViewCount(id);
             return View(viewModel);
         }
 
         #region AJAX Methods
 
-        public async Task<IActionResult> GetCommentByKnowledgeBaseId(int knowledgeBaseId)
+        public async Task<IActionResult> GetCommentsByKnowledgeBaseId(int knowledgeBaseId, int pageIndex = 1, int pageSize = 2)
         {
-            var data = await _knowledgeBaseApiClient.GetCommentsTree(knowledgeBaseId);
+            var data = await _knowledgeBaseApiClient.GetCommentsTree(knowledgeBaseId, pageIndex, pageSize);
+            return Ok(data);
+        }
+
+        public async Task<IActionResult> GetRepliedCommentsByKnowledgeBaseId(int knowledgeBaseId, int rootCommentId, int pageIndex = 1, int pageSize = 2)
+        {
+            var data = await _knowledgeBaseApiClient.GetRepliedComments(knowledgeBaseId, rootCommentId, pageIndex, pageSize);
             return Ok(data);
         }
 
         [HttpPost]
         public async Task<IActionResult> AddNewComment([FromForm] CommentCreateRequest request)
         {
+            if (!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
+            if (!Captcha.ValidateCaptchaCode(request.CaptchaCode, HttpContext))
+            {
+                ModelState.AddModelError("", "Mã xác nhận không đúng");
+                return BadRequest(ModelState);
+            }
+
             var result = await _knowledgeBaseApiClient.PostComment(request);
-            if (result)
-                return Ok();
+            if (result != null)
+                return Ok(result);
             return BadRequest();
+        }
+
+        [HttpPost]
+        public async Task<IActionResult> PostVote([FromForm] VoteCreateRequest request)
+        {
+            var result = await _knowledgeBaseApiClient.PostVote(request);
+            return Ok(result);
+        }
+
+        [HttpPost]
+        public async Task<IActionResult> PostReport([FromForm] ReportCreateRequest request)
+        {
+            if (!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
+            if (!Captcha.ValidateCaptchaCode(request.CaptchaCode, HttpContext))
+            {
+                ModelState.AddModelError("", "Mã xác nhận không đúng");
+                return BadRequest(ModelState);
+            }
+            var result = await _knowledgeBaseApiClient.PostReport(request);
+            return Ok(result);
         }
 
         #endregion AJAX Methods
