@@ -29,7 +29,7 @@ namespace KnowledgeSpace.WebPortal.Services
 
         public async Task<List<T>> GetListAsync<T>(string url, bool requiredLogin = false)
         {
-            var client = _httpClientFactory.CreateClient();
+            var client = _httpClientFactory.CreateClient("BackendApi");
             client.BaseAddress = new Uri(_configuration["BackendApiUrl"]);
             if (requiredLogin)
             {
@@ -44,7 +44,7 @@ namespace KnowledgeSpace.WebPortal.Services
 
         public async Task<T> GetAsync<T>(string url, bool requiredLogin = false)
         {
-            var client = _httpClientFactory.CreateClient();
+            var client = _httpClientFactory.CreateClient("BackendApi");
             client.BaseAddress = new Uri(_configuration["BackendApiUrl"]);
             if (requiredLogin)
             {
@@ -57,13 +57,16 @@ namespace KnowledgeSpace.WebPortal.Services
             return data;
         }
 
-        public async Task<bool> PostAsync<T>(string url, T requestContent, bool requiredLogin = true)
+        public async Task<TResponse> PostAsync<TRequest, TResponse>(string url, TRequest requestContent, bool requiredLogin = true)
         {
-            var client = _httpClientFactory.CreateClient();
+            var client = _httpClientFactory.CreateClient("BackendApi");
             client.BaseAddress = new Uri(_configuration["BackendApiUrl"]);
-
-            var json = JsonConvert.SerializeObject(requestContent);
-            var httpContent = new StringContent(json, Encoding.UTF8, "application/json");
+            StringContent httpContent = null;
+            if (requestContent != null)
+            {
+                var json = JsonConvert.SerializeObject(requestContent);
+                httpContent = new StringContent(json, Encoding.UTF8, "application/json");
+            }
 
             if (requiredLogin)
             {
@@ -71,7 +74,38 @@ namespace KnowledgeSpace.WebPortal.Services
                 client.DefaultRequestHeaders.Authorization = new AuthenticationHeaderValue("Bearer", token);
             }
             var response = await client.PostAsync(url, httpContent);
-            return response.IsSuccessStatusCode;
+            var body = await response.Content.ReadAsStringAsync();
+
+            if (response.IsSuccessStatusCode)
+            {
+                return JsonConvert.DeserializeObject<TResponse>(body);
+            }
+            throw new Exception(body);
+        }
+
+        public async Task<bool> PutAsync<TRequest, TResponse>(string url, TRequest requestContent, bool requiredLogin = true)
+        {
+            var client = _httpClientFactory.CreateClient("BackendApi");
+            client.BaseAddress = new Uri(_configuration["BackendApiUrl"]);
+            HttpContent httpContent = null;
+            if (requestContent != null)
+            {
+                var json = JsonConvert.SerializeObject(requestContent);
+                httpContent = new StringContent(json, Encoding.UTF8, "application/json");
+            }
+
+            if (requiredLogin)
+            {
+                var token = await _httpContextAccessor.HttpContext.GetTokenAsync("access_token");
+                client.DefaultRequestHeaders.Authorization = new AuthenticationHeaderValue("Bearer", token);
+            }
+            var response = await client.PutAsync(url, httpContent);
+            var body = await response.Content.ReadAsStringAsync();
+
+            if (response.IsSuccessStatusCode)
+                return true;
+
+            throw new Exception(body);
         }
     }
 }
