@@ -1,4 +1,6 @@
-﻿using KnowledgeSpace.ViewModels.Systems;
+﻿using KnowledgeSpace.ViewModels;
+using KnowledgeSpace.ViewModels.Contents;
+using KnowledgeSpace.ViewModels.Systems;
 using Microsoft.AspNetCore.Http;
 using Microsoft.Extensions.Configuration;
 using System;
@@ -21,6 +23,12 @@ namespace KnowledgeSpace.WebPortal.Services
         public async Task<UserVm> GetById(string id)
         {
             return await GetAsync<UserVm>($"/api/users/{id}", true);
+        }
+
+        public async Task<Pagination<KnowledgeBaseQuickVm>> GetKnowledgeBasesByUserId(string userId, int pageIndex, int pageSize)
+        {
+            var apiUrl = $"/api/users/{userId}/knowledgeBases?pageIndex={pageIndex}&pageSize={pageSize}";
+            return await GetAsync<Pagination<KnowledgeBaseQuickVm>>(apiUrl, true);
         }
     }
 }
