@@ -23,19 +23,23 @@
     </article>
 
     <form class="row" action="/account/CreateNewKnowledgeBase" method="post" enctype="multipart/form-data" id="frm_new_kb">
-        <div asp-validation-summary="ModelOnly" class="text-danger"></div>
         <div class="span2">
-            <label asp-for="CategoryId" class="control-label"></label>
+        </div>
+        <div class="span6">
+            <div asp-validation-summary="ModelOnly" class="text-danger"></div>
+        </div>
+        <div class="span2">
+            Danh mục
         </div>
         <div class="span6">
             <select asp-for="CategoryId" required class="required input-xlarge"
-                    asp-items="@(new SelectList(ViewBag.Categories,"Id","Name"))">
+                    asp-items="ViewBag.Categories">
             </select>
             <span asp-validation-for="CategoryId" class="text-danger"></span>
         </div>
 
         <div class="span2">
-            <label asp-for="Title" class="control-label"></label>
+            Tiêu đề
         </div>
         <div class="span6">
             <input asp-for="Title" class="required input-xlarge" />
@@ -43,7 +47,7 @@
         </div>
 
         <div class="span2">
-            <label asp-for="Description" class="control-label"></label>
+            Mô tả
         </div>
         <div class="span6">
             <textarea asp-for="Description" rows="5" class="span6"></textarea>
@@ -51,7 +55,7 @@
         </div>
 
         <div class="span2">
-            <label asp-for="Environment" class="control-label"></label>
+            Môi trường
         </div>
         <div class="span6">
             <textarea asp-for="Environment" id="txt_environment" rows="5" class="required span6"></textarea>
@@ -59,7 +63,7 @@
         </div>
 
         <div class="span2">
-            <label asp-for="Problem" class="control-label"></label>
+            Vấn đề
         </div>
         <div class="span6">
             <textarea asp-for="Problem" rows="5" id="txt_problem" class="required span6"></textarea>
@@ -67,7 +71,7 @@
         </div>
 
         <div class="span2">
-            <label asp-for="StepToReproduce" class="control-label"></label>
+            Cách tái hiện
         </div>
         <div class="span6">
             <textarea asp-for="StepToReproduce" id="txt_step_reproduce" rows="5" class="required span6"></textarea>
@@ -75,7 +79,7 @@
         </div>
 
         <div class="span2">
-            <label asp-for="ErrorMessage" class="control-label"></label>
+            Thông báo lỗi
         </div>
         <div class="span6">
             <input asp-for="ErrorMessage" class="input-xlarge" />
@@ -83,7 +87,7 @@
         </div>
 
         <div class="span2">
-            <label asp-for="Workaround" class="control-label"></label>
+            Cách sửa nhanh
         </div>
         <div class="span6">
             <textarea asp-for="Workaround" id="txt_workaround" rows="5" class="required span6"></textarea>
@@ -91,7 +95,7 @@
         </div>
 
         <div class="span2">
-            <label asp-for="Note" class="control-label"></label>
+            Giải pháp
         </div>
         <div class="span6">
             <textarea asp-for="Note" rows="5" id="txt_note" class="required span6"></textarea>
@@ -99,7 +103,7 @@
         </div>
 
         <div class="span2">
-            <label asp-for="Attachments" class="control-label"></label>
+            File đính kèm
         </div>
         <div class="span6" id="attachment_items">
             <p><input type="file" name="attachments" /></p>
@@ -107,11 +111,14 @@
         </div>
 
         <div class="span6 offset2 bm30">
+            <div class="input-group-prepend">
+                <img id="img-captcha" src="~/get-captcha-image" />
+            </div>
+            <input type="text" name="captchaCode" />
             <input type="submit" name="submit" value="Đăng bài" class="btn btn-inverse">
             <img src="/images/loading.gif" id="contact-loader" alt="Đang xử lý...">
         </div>
 
-        <div class="span6 offset2 error-container"></div>
-        <div class="span8 offset2" id="message-sent"></div>
+        <div class="span6 offset2 error-container" id="message-result"></div>
     </form>
 </div>
