@@ -1,7 +1,10 @@
 ﻿@model ListByTagIdViewModel
 @{
     Layout = "_SubLayout";
-    ViewData["Title"] = Model.LabelVm.Name;
+    string suffixPage = Model.Data.PageIndex > 1 ? " - Trang " + Model.Data.PageIndex : "";
+    ViewData["Title"] = Model.LabelVm.Name + suffixPage;
+    ViewData["Description"] = "Các bài viết liên quan đến thẻ " + Model.LabelVm.Name + suffixPage;
+
 }
 <!-- start of page content -->
 <div class="span8 main-listing">
