@@ -19,7 +19,7 @@
                     <ul>
                         @foreach (var item in Model)
                         {
-                            <li><a href="/cat/@item.SeoAlias-@item.Id">@item.Name</a></li>
+                            <li><a href="/cat/@item.SeoAlias-@item.Id" title="@item.Name">@item.Name</a></li>
                         }
                     </ul>
                 </section>
@@ -33,17 +33,6 @@
                             <li>No Tweets loaded !</li>
                         </ul>
                     </div>
-                    <script src="http://twitterjs.googlecode.com/svn/trunk/src/twitter.min.js" type="text/javascript"></script>
-                    <script type="text/javascript">
-                        getTwitters("twitter_update_list", {
-                            id: "960development",
-                            count: 3,
-                            enableLinks: true,
-                            ignoreReplies: true,
-                            clearContents: true,
-                            template: "%text% <span>%time%</span>"
-                        });
-                    </script>
                 </section>
             </div>
 
