@@ -8,7 +8,7 @@
                 <ul class="sub-menu">
                     @foreach (var item in Model)
                     {
-                        <li><a href="/cat/@item.SeoAlias-@item.Id">@item.Name</a></li>
+                        <li><a href="/cat/@item.SeoAlias-@item.Id" title="@Html.Raw(item.Name)">@Html.Raw(item.Name)</a></li>
                     }
                 </ul>
             </li>
