@@ -11,13 +11,14 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
 
     <title>@ViewData["Title"]</title>
+    <meta name="description" content="@ViewData["Description"]">
 
     <link rel="shortcut icon" href="/images/favicon.png" />
 
     <!-- Google Web Fonts-->
-    <link href='http://fonts.googleapis.com/css?family=Open+Sans:300italic,400italic,600italic,700italic,800italic,400,300,600,700,800' rel='stylesheet' type='text/css'>
-    <link href='http://fonts.googleapis.com/css?family=Montserrat:400,700' rel='stylesheet' type='text/css'>
-    <link href='http://fonts.googleapis.com/css?family=Droid+Serif:400,700,400italic,700italic' rel='stylesheet' type='text/css'>
+    <link href='https://fonts.googleapis.com/css?family=Open+Sans:300italic,400italic,600italic,700italic,800italic,400,300,600,700,800' rel='stylesheet' type='text/css'>
+    <link href='https://fonts.googleapis.com/css?family=Montserrat:400,700' rel='stylesheet' type='text/css'>
+    <link href='https://fonts.googleapis.com/css?family=Droid+Serif:400,700,400italic,700italic' rel='stylesheet' type='text/css'>
 
     <!-- Style Sheet-->
     <link rel="stylesheet" href="/css/style.css" />
@@ -31,6 +32,15 @@
     <!--[if lt IE 9]>
     <script src="js/html5.js"></script>
     <![endif]-->
+    <script src="https://cdn.onesignal.com/sdks/OneSignalSDK.js" async=""></script>
+    <script>
+      var OneSignal = window.OneSignal || [];
+      OneSignal.push(function() {
+        OneSignal.init({
+          appId: "6d8ca07d-4c95-48d0-86a6-16c9d3305c39",
+        });
+      });
+    </script>
 </head>
 
 <body>
@@ -40,8 +50,8 @@
             <div class="container">
                 <div class="logo-container">
                     <!-- Website Logo -->
-                    <a href="/" title="Knowledge Base Theme">
-                        <img src="/images/logo.png" alt="Knowledge Base Theme">
+                    <a href="/" title="Knowledge Space">
+                        <img src="/images/logo.png" alt="Knowledge Space logo">
                     </a>
                     <span class="tag-line">Nơi tụ họp giải pháp</span>
                 </div>
@@ -69,14 +79,14 @@
 
     <!-- script -->
     <script type='text/javascript' src='/js/jquery-1.8.3.min.js'></script>
+    <script type='text/javascript' src="/lib/twitter-bootstrap/bootstrap.min.js"></script>
     <script type='text/javascript' src='/js/jquery.easing.1.3.js'></script>
     <script type='text/javascript' src='/js/prettyphoto/jquery.prettyPhoto.js'></script>
-    <script type='text/javascript' src='/js/jflickrfeed.js'></script>
     <script type='text/javascript' src='/js/jquery.liveSearch.js'></script>
     <script type='text/javascript' src='/js/jquery.form.js'></script>
     <script type='text/javascript' src='/js/jquery.validate.min.js'></script>
-    <script type='text/javascript' src="/js/jquery-twitterFetcher.js"></script>
     <script type='text/javascript' src='/js/custom.js'></script>
+    <script type="text/javascript" src="/js/site.js"></script>
     <partial name="_ValidationScriptsPartial" />
     @RenderSection("scripts", required: false)
 </body>
