@@ -9,7 +9,6 @@ using Moq;
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Text;
 using System.Threading.Tasks;
 using Xunit;
 
@@ -47,7 +46,7 @@ namespace KnowledgeSpace.BackendServer.UnitTest.Controllers
         {
             _mockRoleManager.Setup(x => x.CreateAsync(It.IsAny<IdentityRole>()))
                 .ReturnsAsync(IdentityResult.Success);
-            var rolesController = new RolesController(_mockRoleManager.Object,_context);
+            var rolesController = new RolesController(_mockRoleManager.Object, _context);
             var result = await rolesController.PostRole(new RoleCreateRequest()
             {
                 Id = "test",
@@ -63,7 +62,7 @@ namespace KnowledgeSpace.BackendServer.UnitTest.Controllers
         {
             _mockRoleManager.Setup(x => x.CreateAsync(It.IsAny<IdentityRole>()))
                 .ReturnsAsync(IdentityResult.Failed(new IdentityError[] { }));
-            var rolesController = new RolesController(_mockRoleManager.Object,_context);
+            var rolesController = new RolesController(_mockRoleManager.Object, _context);
             var result = await rolesController.PostRole(new RoleCreateRequest()
             {
                 Id = "test",
@@ -79,7 +78,7 @@ namespace KnowledgeSpace.BackendServer.UnitTest.Controllers
         {
             _mockRoleManager.Setup(x => x.Roles)
                 .Returns(_roleSources.AsQueryable().BuildMock().Object);
-            var rolesController = new RolesController(_mockRoleManager.Object,_context);
+            var rolesController = new RolesController(_mockRoleManager.Object, _context);
             var result = await rolesController.GetRoles();
             var okResult = result as OkObjectResult;
             var roleVms = okResult.Value as IEnumerable<RoleVm>;
@@ -91,7 +90,7 @@ namespace KnowledgeSpace.BackendServer.UnitTest.Controllers
         {
             _mockRoleManager.Setup(x => x.Roles).Throws<Exception>();
 
-            var rolesController = new RolesController(_mockRoleManager.Object,_context);
+            var rolesController = new RolesController(_mockRoleManager.Object, _context);
 
             await Assert.ThrowsAnyAsync<Exception>(async () => await rolesController.GetRoles());
         }
@@ -102,7 +101,7 @@ namespace KnowledgeSpace.BackendServer.UnitTest.Controllers
             _mockRoleManager.Setup(x => x.Roles)
                 .Returns(_roleSources.AsQueryable().BuildMock().Object);
 
-            var rolesController = new RolesController(_mockRoleManager.Object,_context);
+            var rolesController = new RolesController(_mockRoleManager.Object, _context);
             var result = await rolesController.GetRolesPaging(null, 1, 2);
             var okResult = result as OkObjectResult;
             var roleVms = okResult.Value as Pagination<RoleVm>;
@@ -116,7 +115,7 @@ namespace KnowledgeSpace.BackendServer.UnitTest.Controllers
             _mockRoleManager.Setup(x => x.Roles)
                 .Returns(_roleSources.AsQueryable().BuildMock().Object);
 
-            var rolesController = new RolesController(_mockRoleManager.Object,_context);
+            var rolesController = new RolesController(_mockRoleManager.Object, _context);
             var result = await rolesController.GetRolesPaging("test3", 1, 2);
             var okResult = result as OkObjectResult;
             var roleVms = okResult.Value as Pagination<RoleVm>;
@@ -129,7 +128,7 @@ namespace KnowledgeSpace.BackendServer.UnitTest.Controllers
         {
             _mockRoleManager.Setup(x => x.Roles).Throws<Exception>();
 
-            var rolesController = new RolesController(_mockRoleManager.Object,_context);
+            var rolesController = new RolesController(_mockRoleManager.Object, _context);
 
             await Assert.ThrowsAnyAsync<Exception>(async () => await rolesController.GetRolesPaging(null, 1, 1));
         }
@@ -143,7 +142,7 @@ namespace KnowledgeSpace.BackendServer.UnitTest.Controllers
                     Id = "test1",
                     Name = "test1"
                 });
-            var rolesController = new RolesController(_mockRoleManager.Object,_context);
+            var rolesController = new RolesController(_mockRoleManager.Object, _context);
             var result = await rolesController.GetById("test1");
             var okResult = result as OkObjectResult;
             Assert.NotNull(okResult);
@@ -158,7 +157,7 @@ namespace KnowledgeSpace.BackendServer.UnitTest.Controllers
         {
             _mockRoleManager.Setup(x => x.FindByIdAsync(It.IsAny<string>())).Throws<Exception>();
 
-            var rolesController = new RolesController(_mockRoleManager.Object,_context);
+            var rolesController = new RolesController(_mockRoleManager.Object, _context);
 
             await Assert.ThrowsAnyAsync<Exception>(async () => await rolesController.GetById("test1"));
         }
@@ -175,7 +174,7 @@ namespace KnowledgeSpace.BackendServer.UnitTest.Controllers
 
             _mockRoleManager.Setup(x => x.UpdateAsync(It.IsAny<IdentityRole>()))
                 .ReturnsAsync(IdentityResult.Success);
-            var rolesController = new RolesController(_mockRoleManager.Object,_context);
+            var rolesController = new RolesController(_mockRoleManager.Object, _context);
             var result = await rolesController.PutRole("test", new RoleCreateRequest()
             {
                 Id = "test",
@@ -199,7 +198,7 @@ namespace KnowledgeSpace.BackendServer.UnitTest.Controllers
             _mockRoleManager.Setup(x => x.UpdateAsync(It.IsAny<IdentityRole>()))
                 .ReturnsAsync(IdentityResult.Failed(new IdentityError[] { }));
 
-            var rolesController = new RolesController(_mockRoleManager.Object,_context);
+            var rolesController = new RolesController(_mockRoleManager.Object, _context);
             var result = await rolesController.PutRole("test", new RoleCreateRequest()
             {
                 Id = "test",
@@ -222,7 +221,7 @@ namespace KnowledgeSpace.BackendServer.UnitTest.Controllers
 
             _mockRoleManager.Setup(x => x.DeleteAsync(It.IsAny<IdentityRole>()))
                 .ReturnsAsync(IdentityResult.Success);
-            var rolesController = new RolesController(_mockRoleManager.Object,_context);
+            var rolesController = new RolesController(_mockRoleManager.Object, _context);
             var result = await rolesController.DeleteRole("test");
             Assert.IsType<OkObjectResult>(result);
         }
@@ -240,7 +239,7 @@ namespace KnowledgeSpace.BackendServer.UnitTest.Controllers
             _mockRoleManager.Setup(x => x.DeleteAsync(It.IsAny<IdentityRole>()))
                 .ReturnsAsync(IdentityResult.Failed(new IdentityError[] { }));
 
-            var rolesController = new RolesController(_mockRoleManager.Object,_context);
+            var rolesController = new RolesController(_mockRoleManager.Object, _context);
             var result = await rolesController.DeleteRole("test");
             Assert.IsType<BadRequestObjectResult>(result);
         }
