@@ -1,20 +1,21 @@
 ﻿using KnowledgeSpace.BackendServer.Data;
 using Microsoft.EntityFrameworkCore;
-using System;
-using System.Collections.Generic;
-using System.Text;
 
 namespace KnowledgeSpace.BackendServer.UnitTest
 {
     public class InMemoryDbContextFactory
     {
-        public ApplicationDbContext GetApplicationDbContext()
+        public ApplicationDbContext GetApplicationDbContext(string databaseName = "InMemoryApplicationDatabase")
         {
             var options = new DbContextOptionsBuilder<ApplicationDbContext>()
-                       .UseInMemoryDatabase(databaseName: "InMemoryApplicationDatabase")
+                       .UseInMemoryDatabase(databaseName: databaseName)
                        .Options;
             var dbContext = new ApplicationDbContext(options);
-
+            if (dbContext != null)
+            {
+                dbContext.Database.EnsureDeleted();
+                dbContext.Database.EnsureCreated();
+            }
             return dbContext;
         }
     }
